+++ server/lib/permaplanner_web/router.ex
@@ -45,7 +45,14 @@ defmodule PermaplannerWeb.Router do
     pipe_through [:api, :require_authenticated, :require_totp_confirmed]
 
     resources "/gardens", GardenController, only: [:index, :create, :show, :update, :delete]
+    get "/gardens/:garden_id/shares", GardenShareController, :index
+    post "/gardens/:garden_id/shares", GardenShareController, :create
+    delete "/gardens/:garden_id/shares/:id", GardenShareController, :delete
     post "/legacy-import/local", LegacyImportController, :local
+  end
+
+  scope "/", PermaplannerWeb do
+    get "/share/:id", GardenShareController, :show
   end
 
   scope "/", PermaplannerWeb do